+++ pages/seller/event/new.js
@@ -10,8 +10,12 @@ import Radio from '@components/input/Radio';
 import * as btn from '@components/input/Button';
 import ImageUploader from '@components/input/ImageUploader';
 import SellerLayout from '@components/seller/SellerLayout';
-import { numberToMonetary } from '@utils/moneyUtil';
-import { getTimeNotKor, getState, isEmpty } from '@utils/functions';
+import {
+  getTimeNotKor,
+  getState,
+  isEmpty,
+  numberToMonetary,
+} from '@utils/functions';
 import { EVENT_TYPE, PRIZE_TYPE } from '@utils/constants/types';
 import { EVENT_ERROR } from '@utils/constants/errors';
