+++ pages/user/order/index.js
@@ -4,13 +4,12 @@ import styled from '@emotion/styled';
 import { createContext, useEffect, useState } from 'react';
 import Input from '@components/input/Input';
 import ProductList from '@components/order/ProductList';
-import { numberToMonetary } from '@utils/moneyUtil';
 import PayList from '@components/order/PayList';
 import { POST, GET } from '@apis/defaultApi';
 import { ThemeGray2 } from '@utils/constants/themeColor';
 import { Blue } from '@components/input/Button';
 import { useRouter } from 'next/router';
-import { isEmpty } from '@utils/functions';
+import { isEmpty, numberToMonetary } from '@utils/functions';
 
 export const AppContext = createContext();
