+++ utils/functions.js
@@ -65,3 +65,10 @@ export var isEmpty = function (value) {
     return false;
   }
 };
+
+export function numberToMonetary(number) {
+  if (!number) return;
+  const numCheck = /^[0-9,]/.test(number);
+  if (!numCheck && number) return;
+  return number.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
+}
